feat: Add cache to filtered data
